monitor/requirements.txt
kafka-python==2.0.6
python-dotenv==1.0.1
pytest==8.3.5
pytest-cov==6.0.0
requests==2.32.3
